backend/requirements.txt
django
djangorestframework
django-environ
mysqlclient